requirements.txt
Django==3.1
mysqlclient==2.0.1
djangorestframework==3.11.1